requirements.txt
django>=4.2.0
fastapi>=0.68.0
uvicorn>=0.15.0
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
redis>=4.0.0
gunicorn>=20.1.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
pytest>=8.3.4
pytest-django>=4.9.0
pytest-asyncio>=0.25.0
httpx>=0.24.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0
coverage>=6.0.0 